+++ CHANGELOG.md
@@ -14,6 +14,7 @@ FIXED:
 - Empty version information appeared in jsonApi object
 - Constructor of `ToOneRelationships` and `ToManyRelationships` were messed up
 - Getters in `MediaTypeUnacceptable` and `MediaTypeUnsupported` didn't return the media type name
+- Pagination objects are correctly instantiated from query parameters
 
 ## 0.5.0 - 2015-09-11